_requirements.txt
ccxt
pandas
numpy
ta
streamlit
pyyaml
plotly
requests
